Add separate component to display data on StatsBar
